Implemented token expiration logic. Crated utils.js for server.
